// apps/shared/src/components/Tree/types.ts
/**
 * Copyright (C) 2025 Monadfix OÜ
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */

export type ItemId = string

export interface TreeData<T extends TreeItem = TreeItem> {
  rootId: ItemId
  items: Record<ItemId, T>
}

export type TreeItemData = any

export type TreeItem<T = {}> = {
  id: ItemId
  children: ItemId[]
  hasChildren?: boolean
  isExpanded?: boolean
  isChildrenLoading?: boolean
  data?: TreeItemData
} & T

export type FlattenedTree = FlattenedItem[]

export type Path = number[]

export type FlattenedItem = {
  item: TreeItem
  path: Path
}

export type TreeSourcePosition = {
  parentId: ItemId
  index: number
}

export type TreeDestinationPosition = {
  parentId: ItemId
  index?: number
}